MP1/6DOF/requirements.txt
absl-py
torch
torchvision
numpy
Pillow
tqdm
tensorboardX
tensorboard
